requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
facenet-pytorch>=2.5.0
gradio>=3.0.0
numpy>=1.21.0
matplotlib>=3.3.0 